refactor: hook_port infrastructure/ports/ → core/orchestration/hook_port.py
ports-migrate: 14 소비자 갱신. 행동 변경 없음.
