+++ core/automation/snapshot.py
@@ -19,7 +19,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from core.infrastructure.ports.hook_port import HookSystemPort
+    from core.orchestration.hook_port import HookSystemPort
 
 log = logging.getLogger(__name__)
