+++ core/cli/__init__.py
@@ -81,10 +81,10 @@ from core.cli.tool_handlers import (
 from core.cli.ui.console import console
 from core.cli.ui.status import GeodeStatus
 from core.config import settings
-from core.infrastructure.ports.hook_port import HookSystemPort
 from core.llm.commentary import (
     generate_commentary,
 )
+from core.orchestration.hook_port import HookSystemPort
 
 log = logging.getLogger(__name__)
