+++ core/orchestration/bootstrap.py
@@ -12,7 +12,7 @@ from collections.abc import Callable
 from dataclasses import dataclass, field
 from typing import Any
 
-from core.infrastructure.ports.hook_port import HookSystemPort
+from core.orchestration.hook_port import HookSystemPort
 from core.orchestration.hooks import HookEvent
 
 log = logging.getLogger(__name__)
