+++ core/orchestration/isolated_execution.py
@@ -19,7 +19,7 @@ from dataclasses import dataclass, field
 from enum import Enum
 from typing import Any
 
-from core.infrastructure.ports.hook_port import HookSystemPort
+from core.orchestration.hook_port import HookSystemPort
 from core.orchestration.hooks import HookEvent
 
 log = logging.getLogger(__name__)
